refactor: 시도 횟수 VO 분리
시도 횟수 VO 구현,
시도 횟수 검증 기능 구현,
getter 구현
