+++ src/main/java/racingcar/racinggame/RacingGame.java
@@ -4,17 +4,14 @@ import java.util.List;
 
 import racingcar.racinggame.io.input.InputHandler;
 import racingcar.racinggame.io.output.OutputHandler;
+import racingcar.racinggame.race.TryCount;
 import racingcar.racinggame.race.car.Cars;
 import racingcar.util.RandomNumberGenerator;
 
 public class RacingGame {
 
-	private static final int MIN_TRY_COUNT = 1;
-	private static final int MAX_TRY_COUNT = 10_000;
 	private static final int MIN_RANDOM_NUMBER = 0;
 	private static final int MAX_RANDOM_NUMBER = 9;
-	private static final String OVER_FLOW_TRY_COUNT_RANGE = "시도 횟수가 지정된 범위를 벗어났습니다.";
-
 	private final InputHandler inputHandler;
 	private final OutputHandler outputHandler;
 
@@ -25,7 +22,7 @@ public class RacingGame {
 
 	public void run() {
 		Cars cars = createCarsFromUserInput();
-		int tryCount = getTryCountFromUerInput();
+		TryCount tryCount = getTryCountFromUerInput();
 
 		raceAllRound(cars, tryCount);
 		notifyWinner(cars);
@@ -37,26 +34,18 @@ public class RacingGame {
 		return Cars.create(carNames);
 	}
 
-	private int getTryCountFromUerInput() {
+	private TryCount getTryCountFromUerInput() {
 		outputHandler.showTryCountInputComment();
 
 		int tryCount = inputHandler.getTryCount();
-		validateTryCountRange(tryCount);
-
-		return tryCount;
+		return TryCount.of(tryCount);
 	}
 
-	private void validateTryCountRange(int tryCount) {
-		if (tryCount < MIN_TRY_COUNT || MAX_TRY_COUNT < tryCount) {
-			throw new IllegalArgumentException(OVER_FLOW_TRY_COUNT_RANGE);
-		}
-	}
-
-	private void raceAllRound(Cars cars, int tryCount) {
+	private void raceAllRound(Cars cars, TryCount tryCount) {
 		outputHandler.showResultComment();
 
 		int carCount = cars.getCarCount();
-		for (int i = 0; i < tryCount; i++) {
+		for (int i = 0; i < tryCount.getCount(); i++) {
 			raceEachRound(cars, carCount);
 		}
 	}
